requirements.txt
flask
flask-classy
flask-cors
requests
python-dotenv
docker
pymongo
checksumdir
kafka-python
bs4
feed-utils==0.6.4
pyjwt
